Adding UI elements
This commit adds all the UI elements of this simple app: a number of labels
and a button that is invisible by default.

The biggest part of this commit is a widget called 'display'. We use this widget
to display all kinds of information, from the module status to the reported
heart rate measurement.