Connecting to the heart rate monitor
This commit adds the code that connects the app to the
heart rate monitor peripheral.

When a heart rate monitor is discovered, scanning is stopped
and a connection attempt is made. When the connection
succeeds ('connectedPeripheral' event is received),
the app stores the peripheral object representing the
now connected peripheral and updates the user interface.