Discovering characteristics
This commit adds the code that discovers and handles
characteristics of interest -- that is, the characteristic
used to report heart rate measurements, and the
characteristic used to report battery status.

The characteristic used to report heart rate measurements
is one that the app can subscribe to, so we do that.
This will automatically result in receiving
updatedValueForCharacteristic events whenever the
heart rate monitor reports a new heart rate value.
We will handle these updates in the next commit.

The battery status characteristic can be polled on
demand -- that's why we the "Check Battery Status"
button. We will add the polling later. For now, we will
just save the object that represent this characteristic
in the batteryChar variable.