Handling connection failures and lost connections
This commit adds two event handlers that deal with
the loss of connection. It is essential that the app
handles these events gracefully, because
BLE communication, like any communication with
a remote entity, can be interrupted at any time.

In this case, the most important task is to restart
scanning, so that the device can be rediscovered
and connection can be established again.