Handling heart rate measurement updates
The heart rate monitor sends us periodic updates
about the measured heart rate. The BLE module
sends 'updatedValueForCharacteristic' events
when this happens. We verify that the update is
for the heart rate measurement characteristic,
and translate it to a 'heartRateUpdate' event,
and in the handler of that event update the
user interface.